requirements.txt
Django==4.0.3
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2